requirements.txt
pillow
PyYAML
numpy>=1.22
scipy==1.5.4
scikit-learn==0.24.2
opencv-python==4.5.5.64

